added jws, dealing with errors

+++ mern-app/backend/controllers/posts.js
@@ -7,13 +7,42 @@ NOTE: Remember that all routes on this page are prefixed with `localhost:3000/ap
 
 /* Require modules
 --------------------------------------------------------------- */
-const express = require('express')
-const router = express.Router()
+const jwt = require('jwt-simple');
+const express = require('express');
+const router = express.Router();
 
 
 /* Require the db connection, and models
 --------------------------------------------------------------- */
 const db = require('../models')
+
+
+/* Require the JWT config
+--------------------------------------------------------------- */
+const config = require('../../jwt.config.js')
+
+
+/* Middleware that checks if a JWT sent from the client is valid.
+   Used for all routes that require authorization
+--------------------------------------------------------------- */
+const authMiddleware = (req, res, next) => {
+    // Check if the 'Authorization' header is present and has the token
+    const token = req.headers.authorization;
+    if (token) {
+        try {
+            // Decode the token using the secret key and add the decoded payload to the request object
+            const decodedToken = jwt.decode(token, config.jwtSecret);
+            req.user = decodedToken;
+            next();
+        } catch (err) {
+            // Return an error if the token is invalid
+            res.status(401).json({ message: 'Invalid token' });
+        }
+    } else {
+        // Return an error if the 'Authorization' header is missing or has the wrong format
+        res.status(401).json({ message: 'Missing or invalid Authorization header' });
+    }
+};
 
 
 /* Routes
@@ -25,8 +54,11 @@ router.get('/', function (req, res) {
 })
 
 // Create Route (POST/Create)
-router.post('/', (req, res) => {
-    db.Post.create(req.body)
+router.post('/', authMiddleware, (req, res) => {
+    db.Post.create({
+        ...req.body,
+        userID: req.user.id
+    })
         .then(post => res.json(post))
 })
 
@@ -37,21 +69,54 @@ router.get('/:id', function (req, res) {
 })
 
 // Update Route (PUT/Update)
-router.put('/:id', (req, res) => {
-    db.Post.findByIdAndUpdate(
-        req.params.id,
-        req.body,
-        { new: true }
-    )
-        .then(post => res.json(post))
+// router.put('/:id', (req, res) => {
+//     db.Post.findByIdAndUpdate(
+//         req.params.id,
+//         req.body,
+//         { new: true }
+//     )
+//         .then(post => res.json(post))
+// })
+
+router.put('/:id', authMiddleware, async (req, res) => {
+    // Check if the user who sent the update request is the same user who created the comment
+    const userContent = await db.Post.findById(req.params.id)
+    if (userContent.userId === req.user.id) {
+        // If it is the original author, update the comment
+        const newContent = await db.Post.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            { new: true }
+        )
+        res.json(newContent)
+    } else {
+        res.status(401).json({ message: 'Invalid user or token' });
+    }
 })
+
 
 // Destroy Route (DELETE/Delete)
-router.delete('/:id', (req, res) => {
-    db.Post.findByIdAndRemove(req.params.id)
-        .then(() => res.send('You deleted post ' + req.params.id))
+// router.delete('/:id', authMiddleware, (req, res) => {
+//     db.Post.findByIdAndRemove(req.params.id)
+//         .then(() => res.send('You deleted post ' + req.params.id))
+// })
+
+router.delete('/:id', authMiddleware, async (req, res) => {
+    // Check if the user who sent the delete request is the same user who created the comment
+    const userContent = await db.Post.findById(req.params.id)
+    if (userContent.userId === req.user.id) {
+        const deletedContent = await db.Post.findByIdAndRemove(req.params.id)
+        res.send('You deleted comment ' + deletedContent._id)
+    } else {
+        res.status(401).json({ message: 'Invalid user or token' });
+    }
 })
 
+// ISSUES ON EDIT/DELETE ROUTES:
+// {
+//     "message": "Invalid user or token"
+// }
+// seems like I may be entering the wrong token information?? not sure 
 
 /* Export these routes so that they are accessible in `server.js`
 --------------------------------------------------------------- */
